Bumps x264 to r999, bringing with it, among other things, the removal of the b-rdo and bime options and new subme modes to replace them.
Subme 5 or greater now includes bime.
Subme 7 is now subme 6 with b-rdo
Subme 8 is the old subme 7 with b-rdo
Subme 9 is new, does RD-refinement for b-frames.

MacGui presets and advanced tab have been updated to reflect these changes (don't forget to update built-in presets).
CLI will wait, as manicure.rb is in a state of flux.
WinGui and LinGui still need to be updated.

git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@1821 b64f7644-9d1e-0410-96f1-a4d463321fa5