fix a couple transport stream issues
when the first packet seen is audio, the stream timing for the audio
must be initialized a little different then when it follows video

add a more thorough check for duplicate packets.
a ts stream that is the result of splicing multiple clips together can
have duplicate continuity count values.  usually this means that a duplicate
packet exists and the duplicate is dropped.  but in the case of spliced
clips, the packet should not be dropped.



git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@2895 b64f7644-9d1e-0410-96f1-a4d463321fa5