libhb: Auto Passthru and audio/video encoder lists.
This adds Auto Passthru as another encoder in libhb.

Used in LinGUI and CLI. MacGUI features a partial impplmentation (with hardcoded values), which is disabled for now.

Also, audio and video encoders are now stored in "lists", like we previously did for mixdowns.

New mixdown "None" (HB_AMIXDOWN_NONE) added to hb_audio_mixdowns (was previously declared in MacGUI/LinGUI).



git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@4237 b64f7644-9d1e-0410-96f1-a4d463321fa5