CLI: add loose cropping option
Adds --loose-crop option.  Loose croping rounds up (or down) the crop
values to the specified mod so that scaling is not required.  One
noteworthy efect of this is that both display aspect and pixel aspect in
the output will be exactly the same as the source when anamorphic is
enabled.

git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@4255 b64f7644-9d1e-0410-96f1-a4d463321fa5